Rewrote the preloaded models to be based on the new profile structure. Full test suite now passes. Shockingly, the docs also build, though the docs are now grossly out of date. Many tests have been commented out, some of which need to come back.
Changes to be committed:
	modified:   halotools/empirical_models/__init__.py
	deleted:    halotools/empirical_models/alt_hod_model_factory.py
	deleted:    halotools/empirical_models/alt_mock_factories.py
	modified:   halotools/empirical_models/conc_mass_models.py
	deleted:    halotools/empirical_models/gal_prof_factory.py
	deleted:    halotools/empirical_models/halo_prof_components.py
	modified:   halotools/empirical_models/mock_factories.py
	modified:   halotools/empirical_models/model_factories.py
	modified:   halotools/empirical_models/preloaded_hod_blueprints.py
	modified:   halotools/empirical_models/preloaded_models.py
	deleted:    halotools/empirical_models/test_empirical_models/test_gal_prof_factory.py
	modified:   halotools/empirical_models/test_empirical_models/test_halo_prof_components.py
	modified:   halotools/empirical_models/test_empirical_models/test_halo_prof_param_components.py
	modified:   halotools/empirical_models/test_empirical_models/test_hod_components.py
	modified:   halotools/empirical_models/test_empirical_models/test_hod_factory.py
	modified:   halotools/empirical_models/test_empirical_models/test_mock_factories.py
	modified:   halotools/empirical_models/test_empirical_models/test_preloaded_hod_blueprints.py
	modified:   halotools/sim_manager/generate_random_sim.py
