+++ halotools/empirical_models/__init__.py
@@ -9,10 +9,7 @@ from .preloaded_models import *
 from .preloaded_hod_blueprints import *
 from .preloaded_subhalo_model_blueprints import *
 from .hod_components import *
-from .halo_prof_param_components import *
-from .halo_prof_components import *
 from .model_defaults import *
-from .gal_prof_factory import *
 from .smhm_components import *
 from .sfr_components import *
 from .abunmatch import *
