+++ halotools/empirical_models/conc_mass_models.py
@@ -1,12 +1,8 @@
 # -*- coding: utf-8 -*-
 """
 
-halotools.halo_prof_param_components contains the classes and functions 
-defining relations between dark matter halos and the parameters 
-governing their internal structure. The classic example of such a component 
-is a relation between halo mass and NFW concentration. More generally, 
-a halo profile parameter model is just a mapping between *any* 
-halo property and *any* profile parameter. 
+Classes and functions defining relations between NFW concentration 
+and halo mass. 
 
 """
