+++ halotools/empirical_models/mock_factories.py
@@ -10,17 +10,17 @@ Currently only composite HOD models are supported.
 
 import numpy as np
 from multiprocessing import cpu_count
+from copy import copy 
 
 from astropy.extern import six
 from abc import ABCMeta, abstractmethod, abstractproperty
 
 from astropy.table import Table 
 
-from . import model_helpers as model_helpers
-from . import model_defaults
+from . import model_helpers, model_defaults
 from .mock_helpers import three_dim_pos_bundle, infer_mask_from_kwargs
 
-from ..custom_exceptions import HalotoolsError
+from ..custom_exceptions import *
 
 try:
     from .. import mock_observables
@@ -497,7 +497,6 @@ class HodMockFactory(MockFactory):
             If set to ``False``, the class will perform all pre-processing tasks 
             but will not call the ``model`` to populate the ``galaxy_table`` 
             with mock galaxies and their observable properties. Default is ``True``. 
-
         """
 
         super(HodMockFactory, self).__init__(populate=populate, **kwargs)
@@ -532,8 +531,7 @@ class HodMockFactory(MockFactory):
 
         ################ Make cuts on halo catalog ################
         # Select host halos only, since this is an HOD-style model
-        host_halo_cut = (self.halo_table['halo_upid']==-1)
-        self.halo_table = self.halo_table[host_halo_cut]
+        self.halo_table = self.snapshot.host_halos
 
         # make a conservative mvir completeness cut 
         # This can be relaxed by changing sim_defaults.Num_ptcl_requirement
@@ -552,17 +550,7 @@ class HodMockFactory(MockFactory):
                 self.halo_table[new_haloprop_key] = new_haloprop_func(halo_table=self.halo_table)
                 self.additional_haloprops.append(new_haloprop_key)
 
-        # Create new columns for the halo catalog associated with each 
-        # parameter of each halo profile model, e.g., 'NFWmodel_conc'. 
-        # New column names are the keys of the halo_prof_func_dict dictionary; 
-        # new column values are computed by the function objects in halo_prof_func_dict 
-        for halo_prof_param_key in self.model.prof_param_keys:
-            method_name = halo_prof_param_key + '_halos'
-            method_behavior = getattr(self.model, method_name)
-            self.halo_table[halo_prof_param_key] = method_behavior(halo_table=self.halo_table)
-            self.additional_haloprops.append(halo_prof_param_key)
-
-        self.model.build_halo_prof_lookup_tables(**kwargs)
+        self.model.build_lookup_tables(**kwargs)
 
     def populate(self, **kwargs):
         """ Method populating halos with mock galaxies. 
@@ -589,23 +577,10 @@ class HodMockFactory(MockFactory):
                 self.galaxy_table[halocatkey][gal_type_slice] = np.repeat(
                     self.halo_table[halocatkey], self._occupation[gal_type], axis=0)
 
-            # Call the galaxy profile components
-            for prof_param_key in self.model.prof_param_keys:
-                method_name = prof_param_key + '_' + gal_type
-                method_behavior = getattr(self.model, method_name)
-                self.galaxy_table[prof_param_key][gal_type_slice] = (
-                    method_behavior(halo_table = self.galaxy_table[gal_type_slice])
-                    )
-
-            # Assign positions 
-            pos_method_name = 'pos_'+gal_type
-
-            self.galaxy_table['x'][gal_type_slice], \
-            self.galaxy_table['y'][gal_type_slice], \
-            self.galaxy_table['z'][gal_type_slice] = (
-                getattr(self.model, pos_method_name)(
-                    halo_table=self.galaxy_table[gal_type_slice])
-                )
+        for method in self._remaining_methods_to_call:
+            func = getattr(self.model, method)
+            gal_type_slice = self._gal_type_indices[func.gal_type]
+            func(halo_table = self.galaxy_table[gal_type_slice])
                 
         # Positions are now assigned to all populations. 
         # Now enforce the periodic boundary conditions for all populations at once
@@ -632,17 +607,44 @@ class HodMockFactory(MockFactory):
 
         self.galaxy_table = Table() 
 
+        # We will keep track of the calling sequence with a list called _remaining_methods_to_call
+        # Each time a function in this list is called, we will remove that function from the list
+        # Mock generation will be complete when _remaining_methods_to_call is exhausted
+        self._remaining_methods_to_call = copy(self.model._mock_generation_calling_sequence)
+
+        # Call all composite model methods that should be called prior to mc_occupation 
+        # All such function calls must be applied to the halo_table, since we do not yet know 
+        # how much memory we need for the mock galaxy_table
+        galprops_assigned_to_halo_table = []
+        for func_name in self.model._mock_generation_calling_sequence:
+            if 'mc_occupation' in func_name:
+                break
+            else:
+                func = getattr(self.model, func_name)
+                func(halo_table = self.halo_table)
+                galprops_assigned_to_halo_table_by_func = func._galprop_dtypes_to_allocate.names
+                galprops_assigned_to_halo_table.extend(galprops_assigned_to_halo_table_by_func)
+                self._remaining_methods_to_call.remove(func_name)
+        # Now update the list of additional_haloprops, if applicable
+        # This is necessary because each of the above function calls created new 
+        # columns for the *halo_table*, not the *galaxy_table*. So we will need to use 
+        # np.repeat inside mock.populate() so that mock galaxies inherit these newly-created columns
+        # Since there is already a loop over additional_haloprops inside mock.populate() that does this, 
+        # then all we need to do is append to this list
+        galprops_assigned_to_halo_table = list(set(
+            galprops_assigned_to_halo_table))
+        self.additional_haloprops.extend(galprops_assigned_to_halo_table)
+        self.additional_haloprops = list(set(self.additional_haloprops))
+
         self._occupation = {}
         self._total_abundance = {}
         self._gal_type_indices = {}
 
         first_galaxy_index = 0
         for gal_type in self.gal_types:
-            #print("Working on gal_type %s" % gal_type)
-            #
             occupation_func_name = 'mc_occupation_'+gal_type
             occupation_func = getattr(self.model, occupation_func_name)
-            # Call the component model to get a MC 
+            # Call the component model to get a Monte Carlo
             # realization of the abundance of gal_type galaxies
             self._occupation[gal_type] = occupation_func(halo_table=self.halo_table)
 
@@ -656,25 +658,29 @@ class HodMockFactory(MockFactory):
             self._gal_type_indices[gal_type] = slice(
                 first_galaxy_index, last_galaxy_index)
             first_galaxy_index = last_galaxy_index
-
+            # Remove the mc_occupation function from the list of methods to call
+            self._remaining_methods_to_call.remove(occupation_func_name)
+            
         self.Ngals = np.sum(self._total_abundance.values())
 
         # Allocate memory for all additional halo properties, 
-        # including profile parameters of the halos such as 'NFWmodel_conc'
+        # including profile parameters of the halos such as 'conc_NFWmodel'
         for halocatkey in self.additional_haloprops:
             self.galaxy_table[halocatkey] = np.zeros(self.Ngals, 
                 dtype = self.halo_table[halocatkey].dtype)
 
-        # Separately allocate memory for the values of the (possibly biased)
-        # galaxy profile parameters such as 'gal_NFWmodel_conc'
+        # Separately allocate memory for the galaxy profile parameters
         for galcatkey in self.model.prof_param_keys:
-            self.galaxy_table[galcatkey] = np.zeros(self.Ngals, dtype = 'f4')
+            self.galaxy_table[galcatkey] = 0.
 
         self.galaxy_table['gal_type'] = np.zeros(self.Ngals, dtype=object)
 
-        phase_space_keys = ['x', 'y', 'z', 'vx', 'vy', 'vz']
-        for key in phase_space_keys:
-            self.galaxy_table[key] = np.zeros(self.Ngals, dtype = 'f4')
+        dt = self.model._galprop_dtypes_to_allocate
+        for key in dt.names:
+            self.galaxy_table[key] = np.zeros(self.Ngals, dtype = dt[key].type)
+
+
+
 
 
 class SubhaloMockFactory(MockFactory):
