+++ halotools/empirical_models/test_empirical_models/test_hod_components.py
@@ -4,7 +4,6 @@ import numpy as np
 from .. import preloaded_hod_blueprints
 from .. import model_defaults
 from .. import hod_components
-from .. import gal_prof_factory
 
 from astropy.table import Table
 from copy import copy
